wip(#3103): split runtime.ts — extract wasi/string-constants/iterator-polyfills
Pure host-side move; tsc clean, emit-identity byte-identical (39/39). ✓
