+++ src/runtime.ts
@@ -17,7 +17,20 @@
 import { compileSource } from "./compiler.js";
 import type { ImportDescriptor, ImportIntent, ImportPolicy } from "./index.js";
 import { createEvalShim, createNewFunctionShim } from "./runtime-eval.js";
-import { hasLoneSurrogate, hexCodeUnits, STRING_CONSTANTS16_NS } from "./string-surrogate.js";
+import { STRING_CONSTANTS16_NS } from "./string-surrogate.js";
+import {
+  _GeneratorState,
+  _AsyncGeneratorState,
+  _getSynthArrayIteratorPrototype,
+  _getGeneratorInstancePrototype,
+  _getGeneratorFunctionPrototype,
+  _getAsyncGeneratorInstancePrototype,
+  _getAsyncGeneratorFunctionPrototype,
+  _installIteratorHelperPolyfills,
+} from "./runtime/iterator-polyfills.js";
+import { buildStringConstants, buildStringConstants16 } from "./runtime/string-constants.js";
+export { buildStringConstants, buildStringConstants16 };
+export { buildWasiPolyfill } from "./runtime/wasi-polyfill.js";
 
 /**
  * Portable require() for loading Node.js builtin modules (#1044).
@@ -228,488 +241,6 @@ function _getOrVivifyFnPrototype(
 }
 
 /**
- * (#1516) Per-generator-instance state: `{buf, index, pendingThrow}`.
- *
- * Storing state in a WeakMap (keyed by the generator instance) instead of as
- * own closure-captured properties on the instance lets the prototype methods
- * `next`/`return`/`throw` be *shared* on `%GeneratorPrototype%` (spec
- * §27.5.1) and perform the GeneratorValidate this-value check (§27.5.3.2)
- * — `if (!_GeneratorState.has(this)) throw TypeError(...)`. The old
- * implementation attached own methods to every instance which made
- * `Generator.prototype.next.call(non_gen)` succeed (wrong).
- *
- * State shape:
- *   buf: any[]            — eager-yield buffer (filled by the generator body)
- *   index: number         — next read position in `buf`
- *   pendingThrow: any     — exception captured by the generator body, to be
- *                            re-thrown on the first `next()` after the
- *                            buffer is drained (#928)
- *   retVal: any           — the generator's `return` value (#2035). Per
- *                            §27.5.3.3 / §27.5.1.2 the return value belongs
- *                            ONLY to the terminal `{value, done:true}` result;
- *                            it must NOT appear as a yielded (`done:false`)
- *                            element. Surfaced once when the buffer drains.
- *   retDone: boolean      — `true` once the terminal `{retVal, done:true}`
- *                            result has been handed out, so subsequent
- *                            `next()` calls return `{value:undefined,done:true}`
- *   asyncWrap?: boolean   — `true` for async-generator state (so the same
- *                            map can back both `%GeneratorPrototype%` and
- *                            `%AsyncGeneratorPrototype%` methods)
- */
-const _GeneratorState = new WeakMap<
-  object,
-  {
-    buf: any[];
-    index: number;
-    pendingThrow: any;
-    retVal?: any;
-    retDone?: boolean;
-    /** (#3032) LAZY thunk mode: the generator-expression closure (an opaque
-     *  wasm externref) whose deferred invocation produces the buffer. Present
-     *  only until the first `next()` (or cleared without running by
-     *  `return`/`throw` on a not-yet-started generator — spec §27.5.3.2:
-     *  resuming a suspendedStart generator abruptly completes it WITHOUT
-     *  running the body). */
-    thunk?: any;
-    /** (#3032) Runs the thunk via the module's `__call_fn_0` export with the
-     *  `__gen_set_eager` flag held, then adopts the inner eager generator's
-     *  state. Captured at `__create_generator` time (needs `callbackState`). */
-    materialize?: () => void;
-  }
->();
-const _AsyncGeneratorState = new WeakMap<
-  object,
-  { buf: any[]; index: number; pendingThrow: any; retVal?: any; retDone?: boolean }
->();
-
-let _GeneratorPrototypeCache: any = null;
-let _GeneratorFunctionPrototypeCache: any = null;
-let _AsyncGeneratorPrototypeCache: any = null;
-let _AsyncGeneratorFunctionPrototypeCache: any = null;
-// (#1639) `genFn.prototype` — the per-function instance prototype. Per spec
-// §27.3.4 / §27.4.4 it is `OrdinaryObjectCreate(%(Async)GeneratorPrototype%)`,
-// one level below the shared `%(Async)GeneratorPrototype%`. Generator instances
-// inherit from this object, so `Object.getPrototypeOf(instance) === genFn.prototype`.
-let _GeneratorInstancePrototypeCache: any = null;
-let _AsyncGeneratorInstancePrototypeCache: any = null;
-let _IteratorPrototypeCache: any = null;
-let _AsyncIteratorPrototypeCache: any = null;
-// (#3049) Shared `%ArrayIteratorPrototype%` stand-in for iterators the
-// `__iterator` host import SYNTHESIZES over compiled vec structs. §23.1.5.2:
-// array iterators are ObjectCreate(%ArrayIteratorPrototype%), whose
-// [[Prototype]] is %IteratorPrototype%. The old one-level chain made the
-// spec-shaped walk `getPrototypeOf(getPrototypeOf([][Symbol.iterator]()))`
-// (hardcoded by tests + the runner's `Iterator` shim) overshoot the
-// helper-bearing proto onto Object.prototype → every
-// `Iterator.prototype.<helper>` lookup was undefined. One SHARED cached
-// middle object keeps getPrototypeOf identity stable across iterators.
-let _SynthArrayIteratorPrototypeCache: any = null;
-function _getSynthArrayIteratorPrototype(base: any): any {
-  if (_SynthArrayIteratorPrototypeCache) return _SynthArrayIteratorPrototypeCache;
-  const proto = Object.create(base ?? null);
-  _SynthArrayIteratorPrototypeCache = proto;
-  // §23.1.5.2.2 %ArrayIteratorPrototype% [ @@toStringTag ]
-  Object.defineProperty(proto, Symbol.toStringTag, {
-    value: "Array Iterator",
-    writable: false,
-    enumerable: false,
-    configurable: true,
-  });
-  return proto;
-}
-
-/**
- * Install a built-in method on a prototype with spec-mandated descriptor
- * flags. ES2024 §17 specifies that built-in function objects have:
- *   - `length`: { value: N, writable: false, enumerable: false, configurable: true }
- *   - `name`:   { value: "<name>", writable: false, enumerable: false, configurable: true }
- * and that built-in methods on a prototype are installed with
- *   { writable: true, enumerable: false, configurable: true }.
- */
-function _installBuiltinMethod(
-  proto: object,
-  name: string,
-  length: number,
-  impl: (this: any, ...args: any[]) => any,
-): void {
-  // Re-assign `name`/`length` to match spec descriptors. We name the
-  // function via `Object.defineProperty` so it doesn't appear as an
-  // anonymous arrow in stack traces.
-  Object.defineProperty(impl, "length", {
-    value: length,
-    writable: false,
-    enumerable: false,
-    configurable: true,
-  });
-  Object.defineProperty(impl, "name", {
-    value: name,
-    writable: false,
-    enumerable: false,
-    configurable: true,
-  });
-  Object.defineProperty(proto, name, {
-    value: impl,
-    writable: true,
-    enumerable: false,
-    configurable: true,
-  });
-}
-
-/**
- * Install a static helper (e.g. `Iterator.zip`) with spec-correct property
- * descriptors. The function's own `length` and `name` are reset to the
- * spec-mandated values with attributes `{writable:false, enumerable:false,
- * configurable:true}` (§17 — built-in function `length`/`name`), and the
- * property on `target` itself is `{writable:true, enumerable:false,
- * configurable:true}` (§17 default data-property attributes). TS optional
- * params (`options?`) inflate `fn.length`, so we override it explicitly.
- */
-function _installStaticHelper(
-  target: object,
-  name: string,
-  length: number,
-  impl: (this: any, ...args: any[]) => any,
-): void {
-  Object.defineProperty(impl, "length", {
-    value: length,
-    writable: false,
-    enumerable: false,
-    configurable: true,
-  });
-  Object.defineProperty(impl, "name", {
-    value: name,
-    writable: false,
-    enumerable: false,
-    configurable: true,
-  });
-  Object.defineProperty(target, name, {
-    value: impl,
-    writable: true,
-    enumerable: false,
-    configurable: true,
-  });
-}
-
-/**
- * Build `%IteratorPrototype%` (spec §27.1.2). Its sole own property is
- * `[Symbol.iterator]()` which returns `this`. `%GeneratorPrototype%` inherits
- * from it so generators are iterable. (#1639) We build it explicitly rather
- * than borrowing `globalThis.Iterator.prototype`, which may be absent and in
- * any case is not the object test262 walks to via the generator's proto chain.
- */
-function _getIteratorPrototype(): any {
-  if (_IteratorPrototypeCache) return _IteratorPrototypeCache;
-  const proto = Object.create(Object.prototype);
-  _IteratorPrototypeCache = proto;
-  const fn = function (this: any) {
-    return this;
-  };
-  Object.defineProperty(fn, "length", { value: 0, writable: false, enumerable: false, configurable: true });
-  Object.defineProperty(fn, "name", {
-    value: "[Symbol.iterator]",
-    writable: false,
-    enumerable: false,
-    configurable: true,
-  });
-  Object.defineProperty(proto, Symbol.iterator, {
-    value: fn,
-    writable: true,
-    enumerable: false,
-    configurable: true,
-  });
-  return proto;
-}
-
-/**
- * Build `%AsyncIteratorPrototype%` (spec §27.1.3). Its sole own property is
- * `[Symbol.asyncIterator]()` which returns `this`. `%AsyncGeneratorPrototype%`
- * inherits from it. (#1639)
- */
-function _getAsyncIteratorPrototype(): any {
-  if (_AsyncIteratorPrototypeCache) return _AsyncIteratorPrototypeCache;
-  const proto = Object.create(Object.prototype);
-  _AsyncIteratorPrototypeCache = proto;
-  const fn = function (this: any) {
-    return this;
-  };
-  Object.defineProperty(fn, "length", { value: 0, writable: false, enumerable: false, configurable: true });
-  Object.defineProperty(fn, "name", {
-    value: "[Symbol.asyncIterator]",
-    writable: false,
-    enumerable: false,
-    configurable: true,
-  });
-  Object.defineProperty(proto, Symbol.asyncIterator, {
-    value: fn,
-    writable: true,
-    enumerable: false,
-    configurable: true,
-  });
-  return proto;
-}
-
-/** Build `%GeneratorPrototype%` (spec §27.5.1). Idempotent. */
-function _getGeneratorPrototype(): any {
-  if (_GeneratorPrototypeCache) return _GeneratorPrototypeCache;
-  // GeneratorPrototype inherits from %IteratorPrototype% so .map/.filter/etc.
-  // (#1367) resolve via the prototype chain, and test262 reaches
-  // %IteratorPrototype% via getPrototypeOf(getPrototypeOf(g.prototype)). (#1639)
-  const proto = Object.create(_getIteratorPrototype());
-  _GeneratorPrototypeCache = proto;
-
-  _installBuiltinMethod(proto, "next", 1, function (this: any, _value?: any) {
-    const state = _GeneratorState.get(this);
-    if (!state) {
-      throw new TypeError("Generator.prototype.next called on incompatible receiver");
-    }
-    // (#3032) Lazy generator: run the deferred body now (first resume).
-    if (state.materialize) state.materialize();
-    if (state.index < state.buf.length) {
-      return { value: state.buf[state.index++], done: false };
-    }
-    if (state.pendingThrow !== null && state.pendingThrow !== undefined) {
-      const e = state.pendingThrow;
-      state.pendingThrow = null;
-      throw e;
-    }
-    // Buffer drained: surface the generator's `return` value ONCE as the
-    // terminal `{value, done:true}` result (#2035, §27.5.1.2 step 6). After
-    // that, every further `next()` returns `{value:undefined, done:true}`.
-    if (!state.retDone) {
-      state.retDone = true;
-      return { value: state.retVal, done: true };
-    }
-    return { value: undefined, done: true };
-  });
-
-  _installBuiltinMethod(proto, "return", 1, function (this: any, value?: any) {
-    const state = _GeneratorState.get(this);
-    if (!state) {
-      throw new TypeError("Generator.prototype.return called on incompatible receiver");
-    }
-    // Early termination: skip the rest of the buffer AND suppress the
-    // generator's own return value — the caller-supplied `value` becomes the
-    // terminal result (§27.5.3.3). Mark retDone so a later next() is terminal.
-    // (#3032) A not-yet-started lazy generator completes WITHOUT running its
-    // body (§27.5.3.2 GeneratorResumeAbrupt on suspendedStart) — drop the thunk.
-    state.thunk = undefined;
-    state.materialize = undefined;
-    state.index = state.buf.length;
-    state.retDone = true;
-    return { value, done: true };
-  });
-
-  _installBuiltinMethod(proto, "throw", 1, function (this: any, e?: any) {
-    const state = _GeneratorState.get(this);
-    if (!state) {
-      throw new TypeError("Generator.prototype.throw called on incompatible receiver");
-    }
-    // (#3032) See `return` — abrupt resume of suspendedStart never runs the body.
-    state.thunk = undefined;
-    state.materialize = undefined;
-    state.index = state.buf.length;
-    throw e;
-  });
-
-  // `[Symbol.iterator]` returning `this` — generators are their own iterators
-  // (Iterator.prototype already provides this via @@iterator returning this,
-  // but install it explicitly to be robust against missing %IteratorPrototype%
-  // in older runtimes).
-  if (!(Symbol.iterator in proto)) {
-    Object.defineProperty(proto, Symbol.iterator, {
-      value: function (this: any) {
-        return this;
-      },
-      writable: true,
-      enumerable: false,
-      configurable: true,
-    });
-  }
-
-  // Symbol.toStringTag = 'Generator' (spec §27.5.1.5)
-  Object.defineProperty(proto, Symbol.toStringTag, {
-    value: "Generator",
-    writable: false,
-    enumerable: false,
-    configurable: true,
-  });
-
-  // The `constructor` slot points at %Generator% (= %GeneratorFunction.prototype%).
-  // Spec §27.5.1.1 requires a *data* property {writable:false, enumerable:false,
-  // configurable:true} — not an accessor. `_getGeneratorFunctionPrototype` set its
-  // own cache before it called us (so this call returns the in-progress object
-  // without recursing), making the data value safe to install here.
-  Object.defineProperty(proto, "constructor", {
-    value: _getGeneratorFunctionPrototype(),
-    writable: false,
-    enumerable: false,
-    configurable: true,
-  });
-
-  return proto;
-}
-
-/**
- * (#1639) Build the per-`function*` instance prototype `genFn.prototype`
- * (spec §27.3.4): `OrdinaryObjectCreate(%GeneratorPrototype%)`. Generator
- * instances inherit from this object so the spec chain holds:
- *   instance → genFn.prototype → %GeneratorPrototype% → %IteratorPrototype%.
- */
-function _getGeneratorInstancePrototype(): any {
-  if (_GeneratorInstancePrototypeCache) return _GeneratorInstancePrototypeCache;
-  _GeneratorInstancePrototypeCache = Object.create(_getGeneratorPrototype());
-  return _GeneratorInstancePrototypeCache;
-}
-
-/** Build `%GeneratorFunction.prototype%` (= `%Generator%`, spec §27.3.3). */
-function _getGeneratorFunctionPrototype(): any {
-  if (_GeneratorFunctionPrototypeCache) return _GeneratorFunctionPrototypeCache;
-  // %Generator% inherits from %Function.prototype% so `typeof g.constructor === 'function'`.
-  const proto = Object.create(Function.prototype);
-  _GeneratorFunctionPrototypeCache = proto;
-
-  // `prototype` slot = `%GeneratorPrototype%` (writable: false, !enum, configurable: false per spec —
-  //   §27.3.3.3 — though several engines ship it as configurable; configurable is what test262
-  //   verifyProperty defaults check against).
-  Object.defineProperty(proto, "prototype", {
-    value: _getGeneratorPrototype(),
-    writable: false,
-    enumerable: false,
-    configurable: false,
-  });
-
-  Object.defineProperty(proto, Symbol.toStringTag, {
-    value: "GeneratorFunction",
-    writable: false,
-    enumerable: false,
-    configurable: true,
-  });
-
-  return proto;
-}
-
-/** Build `%AsyncGeneratorPrototype%` (spec §27.6.1). */
-function _getAsyncGeneratorPrototype(): any {
-  if (_AsyncGeneratorPrototypeCache) return _AsyncGeneratorPrototypeCache;
-  // Inherits from %AsyncIteratorPrototype% (#1639) — test262 reaches it via
-  // getPrototypeOf(getPrototypeOf(asyncGen.prototype)).
-  const proto = Object.create(_getAsyncIteratorPrototype());
-  _AsyncGeneratorPrototypeCache = proto;
-
-  function mkResult(value: any, done: boolean) {
-    const plain = { value, done };
-    return {
-      value,
-      done,
-      then(res: any, rej: any) {
-        return Promise.resolve(plain).then(res, rej);
-      },
-    };
-  }
-  function mkError(e: any) {
-    return {
-      done: true,
-      value: undefined as any,
-      then(res: any, rej: any) {
-        return Promise.reject(e).then(res, rej);
-      },
-    };
-  }
-
-  _installBuiltinMethod(proto, "next", 1, function (this: any, _value?: any) {
-    const state = _AsyncGeneratorState.get(this);
-    if (!state) {
-      return mkError(new TypeError("AsyncGenerator.prototype.next called on incompatible receiver"));
-    }
-    if (state.index < state.buf.length) return mkResult(state.buf[state.index++], false);
-    if (state.pendingThrow !== null && state.pendingThrow !== undefined) {
-      const e = state.pendingThrow;
-      state.pendingThrow = null;
-      return mkError(e);
-    }
-    return mkResult(undefined, true);
-  });
-
-  _installBuiltinMethod(proto, "return", 1, function (this: any, value?: any) {
-    const state = _AsyncGeneratorState.get(this);
-    if (!state) {
-      return mkError(new TypeError("AsyncGenerator.prototype.return called on incompatible receiver"));
-    }
-    state.index = state.buf.length;
-    return mkResult(value, true);
-  });
-
-  _installBuiltinMethod(proto, "throw", 1, function (this: any, e?: any) {
-    const state = _AsyncGeneratorState.get(this);
-    if (!state) {
-      return mkError(new TypeError("AsyncGenerator.prototype.throw called on incompatible receiver"));
-    }
-    state.index = state.buf.length;
-    return mkError(e);
-  });
-
-  if (!(Symbol.asyncIterator in proto)) {
-    Object.defineProperty(proto, Symbol.asyncIterator, {
-      value: function (this: any) {
-        return this;
-      },
-      writable: true,
-      enumerable: false,
-      configurable: true,
-    });
-  }
-
-  Object.defineProperty(proto, Symbol.toStringTag, {
-    value: "AsyncGenerator",
-    writable: false,
-    enumerable: false,
-    configurable: true,
-  });
-
-  // Spec §27.6.1.1 — `constructor` is a data property {writable:false,
-  // enumerable:false, configurable:true} pointing at %AsyncGenerator%.
-  Object.defineProperty(proto, "constructor", {
-    value: _getAsyncGeneratorFunctionPrototype(),
-    writable: false,
-    enumerable: false,
-    configurable: true,
-  });
-
-  return proto;
-}
-
-/**
- * (#1639) Build the per-`async function*` instance prototype `genFn.prototype`
- * (spec §27.4.4): `OrdinaryObjectCreate(%AsyncGeneratorPrototype%)`. The chain
- * is: instance → genFn.prototype → %AsyncGeneratorPrototype% → %AsyncIteratorPrototype%.
- */
-function _getAsyncGeneratorInstancePrototype(): any {
-  if (_AsyncGeneratorInstancePrototypeCache) return _AsyncGeneratorInstancePrototypeCache;
-  _AsyncGeneratorInstancePrototypeCache = Object.create(_getAsyncGeneratorPrototype());
-  return _AsyncGeneratorInstancePrototypeCache;
-}
-
-/** Build `%AsyncGeneratorFunction.prototype%` (= `%AsyncGenerator%`, spec §27.4.3). */
-function _getAsyncGeneratorFunctionPrototype(): any {
-  if (_AsyncGeneratorFunctionPrototypeCache) return _AsyncGeneratorFunctionPrototypeCache;
-  const proto = Object.create(Function.prototype);
-  _AsyncGeneratorFunctionPrototypeCache = proto;
-  Object.defineProperty(proto, "prototype", {
-    value: _getAsyncGeneratorPrototype(),
-    writable: false,
-    enumerable: false,
-    configurable: false,
-  });
-  Object.defineProperty(proto, Symbol.toStringTag, {
-    value: "AsyncGeneratorFunction",
-    writable: false,
-    enumerable: false,
-    configurable: true,
-  });
-  return proto;
-}
-
-/**
  * (#1334) Per-object set of property keys that were explicitly deleted via
  * the `delete` operator. WasmGC structs have a fixed shape — fields can't
  * be removed at runtime — so a successful `delete obj.x` only sets the
@@ -757,819 +288,6 @@ const _wasmPropDescs = new WeakMap<object, Map<string | symbol, number>>();
  * Separate from _wasmStructProps because template literals can't stringify Symbols.
  */
 const _wasmStructAccessors = new WeakMap<object, Map<string | symbol, { get?: Function; set?: Function }>>();
-
-/**
- * #1464 — ES2025 Iterator helper polyfills.
- *
- * `Iterator.zip(iters, opts)`, `Iterator.zipKeyed(iterMap, opts)` and
- * `Iterator.concat(...iters)` are recent ES2025 additions that some hosts
- * (Node ≤24 / older V8) don't ship. They are wired through
- * `__extern_method_call` (since `Iterator` is in `BUILTIN_CLASS_NAMES`),
- * so installing a single polyfill on `globalThis.Iterator` makes every
- * call site work uniformly. The polyfill:
- *   - returns helper iterators that inherit from `%Iterator.prototype%`
- *     so chained `.map / .filter / .toArray` continue working
- *   - validates arguments eagerly and calls `return()` on any
- *     already-opened underlying iterator when one rejects
- *   - implements `mode: "shortest" | "longest" | "strict"` for `zip` per
- *     the proposal text (TC39 stage 4, ES2025)
- *
- * Called from `buildImports` (once, guarded by `_iteratorHelpersInstalled`).
- * Safe to call on a host that already ships the helpers — we only install
- * when the method is missing.
- */
-let _iteratorHelpersInstalled = false;
-const _intrinsicStringIterator = String.prototype[Symbol.iterator];
-function _installIteratorHelperPolyfills(): void {
-  if (_iteratorHelpersInstalled) return;
-  _iteratorHelpersInstalled = true;
-  const compilerIteratorProto = _getIteratorPrototype();
-  let I: any = (globalThis as any).Iterator;
-  if (typeof I !== "function" || typeof I.prototype !== "object" || I.prototype == null) {
-    I = function Iterator(this: any) {
-      throw new TypeError("Iterator is not a constructor");
-    };
-    I.prototype = compilerIteratorProto;
-    Object.defineProperty(globalThis, "Iterator", {
-      value: I,
-      writable: true,
-      enumerable: false,
-      configurable: true,
-    });
-  }
-  const Iproto: any = typeof I.prototype === "object" && I.prototype != null ? I.prototype : compilerIteratorProto;
-
-  if (Iproto !== compilerIteratorProto) {
-    let wouldCycle = false;
-    for (let p = Iproto; p != null; p = Object.getPrototypeOf(p)) {
-      if (p === compilerIteratorProto) {
-        wouldCycle = true;
-        break;
-      }
-    }
-    if (!wouldCycle && Object.getPrototypeOf(compilerIteratorProto) !== Iproto) {
-      Object.setPrototypeOf(compilerIteratorProto, Iproto);
-    }
-  }
-
-  function _isObject(value: any): boolean {
-    return (typeof value === "object" && value !== null) || typeof value === "function";
-  }
-
-  function _getIteratorDirect(iter: any): any {
-    if (!_isObject(iter)) {
-      throw new TypeError("Iterator helper: iterator is not an object");
-    }
-    if (typeof iter.next !== "function") {
-      throw new TypeError("Iterator helper: iterator has no next()");
-    }
-    return iter;
-  }
-
-  function _getIterator(iterable: any): any {
-    if (!_isObject(iterable)) {
-      throw new TypeError("Iterator helper: argument is not iterable");
-    }
-    const method = iterable[Symbol.iterator];
-    if (typeof method !== "function") {
-      throw new TypeError("Iterator helper: argument is not iterable");
-    }
-    return _getIteratorDirect(method.call(iterable));
-  }
-
-  function _iteratorStepValue(iter: any): { done: true; value?: undefined } | { done: false; value: any } {
-    const result = iter.next();
-    if (!_isObject(result)) {
-      throw new TypeError("Iterator result is not an object");
-    }
-    if (result.done) return { done: true };
-    return { value: result.value, done: false };
-  }
-
-  // ES2025 GetIteratorFlattenable — accepts an iterable OR a raw iterator.
-  // `Iterator.from` uses iterate-string-primitives; iterator helper flattening
-  // (e.g. flatMap) uses reject-primitives.
-  function _getFlattenable(obj: any, primitiveHandling: "iterate-string-primitives" | "reject-primitives"): any {
-    if (obj == null) {
-      throw new TypeError("Iterator helper: argument is null or undefined");
-    }
-    const primitive = typeof obj !== "object" && typeof obj !== "function";
-    if (primitive) {
-      if (primitiveHandling === "reject-primitives" || typeof obj !== "string") {
-        throw new TypeError("Iterator helper: argument is not iterable");
-      }
-    }
-    let it: any;
-    const stringLike = typeof obj === "string" || obj instanceof String;
-    let sym = primitive ? Reflect.get(Object(obj), Symbol.iterator, obj) : obj[Symbol.iterator];
-    if (typeof sym !== "function" && stringLike && typeof _intrinsicStringIterator === "function") {
-      sym = _intrinsicStringIterator;
-    }
-    if (typeof sym === "function") {
-      it = sym.call(obj);
-    } else if (typeof obj.next === "function") {
-      it = obj;
-    } else {
-      throw new TypeError("Iterator helper: argument is not iterable");
-    }
-    return _getIteratorDirect(it);
-  }
-
-  // ES2025 GetOptionsObject (iterator-sequencing / joint-iteration proposal):
-  // undefined → fresh null-proto object; Object → returned as-is; any other
-  // value (null, boolean, number, string, symbol, bigint) throws TypeError.
-  function _getOptionsObject(options: any): any {
-    if (options === undefined) return Object.create(null);
-    if (typeof options === "object" && options !== null) return options;
-    if (typeof options === "function") return options;
-    throw new TypeError("Iterator options must be undefined or an object");
-  }
-
-  function _makeHelperIterator(nextFn: () => any, returnFn: (v?: any) => any): any {
-    const obj: any = Object.create(Iproto);
-    obj.next = nextFn;
-    obj.return = returnFn;
-    obj[Symbol.iterator] = function () {
-      return this;
-    };
-    return obj;
-  }
-
-  function _requireIteratorReceiver(receiver: any, name: string): any {
-    if (receiver == null || typeof receiver.next !== "function") {
-      throw new TypeError("Iterator.prototype." + name + " called on non-iterator");
-    }
-    return receiver;
-  }
-
-  function _closeIterator(iter: any): void {
-    try {
-      iter?.return?.();
-    } catch {}
-  }
-
-  function _closeIterators(iters: any[], open?: boolean[], except = -1): void {
-    for (let i = 0; i < iters.length; i++) {
-      if (i === except) continue;
-      if (open && !open[i]) continue;
-      if (open) open[i] = false;
-      _closeIterator(iters[i]);
-    }
-  }
-
-  function _readZipOptions(options: any): { mode: "shortest" | "longest" | "strict"; paddingOption: any } {
-    options = _getOptionsObject(options);
-    let mode = options.mode;
-    if (mode === undefined) mode = "shortest";
-    if (mode !== "shortest" && mode !== "longest" && mode !== "strict") {
-      throw new TypeError("Iterator.zip: invalid mode " + String(mode));
-    }
-    let paddingOption: any = undefined;
-    if (mode === "longest") {
-      paddingOption = options.padding;
-      if (paddingOption !== undefined && !_isObject(paddingOption)) {
-        throw new TypeError("Iterator.zip: padding must be an object");
-      }
-    }
-    return { mode, paddingOption };
-  }
-
-  function _makeIteratorZip(
-    iters: any[],
-    mode: "shortest" | "longest" | "strict",
-    padding: any[],
-    finishResults: (results: any[]) => any,
-  ): any {
-    const open = iters.map(() => true);
-    let exhausted = false;
-
-    return _makeHelperIterator(
-      function next() {
-        if (exhausted || iters.length === 0) return { value: undefined, done: true };
-        const results: any[] = [];
-        for (let i = 0; i < iters.length; i++) {
-          let result: any;
-          if (!open[i]) {
-            result = padding[i];
-          } else {
-            let step: { done: true; value?: undefined } | { done: false; value: any };
-            try {
-              step = _iteratorStepValue(iters[i]);
-            } catch (e) {
-              exhausted = true;
-              open[i] = false;
-              _closeIterators(iters, open);
-              throw e;
-            }
-
-            if (step.done) {
-              open[i] = false;
-              if (mode === "shortest") {
-                exhausted = true;
-                _closeIterators(iters, open);
-                return { value: undefined, done: true };
-              }
-              if (mode === "strict") {
-                exhausted = true;
-                if (i !== 0) {
-                  _closeIterators(iters, open);
-                  throw new TypeError("Iterator.zip strict mode: length mismatch");
-                }
-                for (let k = 1; k < iters.length; k++) {
-                  let openStep: { done: true; value?: undefined } | { done: false; value: any };
-                  try {
-                    openStep = _iteratorStepValue(iters[k]);
-                  } catch (e) {
-                    open[k] = false;
-                    _closeIterators(iters, open);
-                    throw e;
-                  }
-                  if (openStep.done) {
-                    open[k] = false;
-                  } else {
-                    _closeIterators(iters, open);
-                    throw new TypeError("Iterator.zip strict mode: length mismatch");
-                  }
-                }
-                return { value: undefined, done: true };
-              }
-
-              if (!open.some(Boolean)) {
-                exhausted = true;
-                return { value: undefined, done: true };
-              }
-              result = padding[i];
-            } else {
-              result = step.value;
-            }
-          }
-          results.push(result);
-        }
-        return { value: finishResults(results), done: false };
-      },
-      function returnFn() {
-        exhausted = true;
-        _closeIterators(iters, open);
-        return { value: undefined, done: true };
-      },
-    );
-  }
-
-  // (#1320) Always route Iterator.from through the bridge implementation. Host
-  // natives cannot call compiled accessor closures installed on primitive
-  // prototypes reliably after the Wasm closure crosses externref.
-  _installStaticHelper(I, "from", 1, function from(iterable: any) {
-    return _getFlattenable(iterable, "iterate-string-primitives");
-  });
-
-  if (typeof Iproto.map !== "function") {
-    _installBuiltinMethod(Iproto, "map", 1, function (this: any, mapper: any) {
-      const outer = _requireIteratorReceiver(this, "map");
-      if (typeof mapper !== "function") {
-        throw new TypeError("Iterator.prototype.map: mapper is not a function");
-      }
-      let counter = 0;
-      let done = false;
-      return _makeHelperIterator(
-        function next() {
-          if (done) return { value: undefined, done: true };
-          const r = outer.next();
-          if (r && r.done) {
-            done = true;
-            return { value: undefined, done: true };
-          }
-          try {
-            return { value: mapper(r.value, counter++), done: false };
-          } catch (e) {
-            done = true;
-            _closeIterator(outer);
-            throw e;
-          }
-        },
-        function returnFn() {
-          done = true;
-          _closeIterator(outer);
-          return { value: undefined, done: true };
-        },
-      );
-    });
-  }
-
-  if (typeof Iproto.filter !== "function") {
-    _installBuiltinMethod(Iproto, "filter", 1, function (this: any, predicate: any) {
-      const outer = _requireIteratorReceiver(this, "filter");
-      if (typeof predicate !== "function") {
-        throw new TypeError("Iterator.prototype.filter: predicate is not a function");
-      }
-      let counter = 0;
-      let done = false;
-      return _makeHelperIterator(
-        function next() {
-          if (done) return { value: undefined, done: true };
-          while (true) {
-            const r = outer.next();
-            if (r && r.done) {
-              done = true;
-              return { value: undefined, done: true };
-            }
-            let keep: any;
-            try {
-              keep = predicate(r.value, counter++);
-            } catch (e) {
-              done = true;
-              _closeIterator(outer);
-              throw e;
-            }
-            if (keep) return { value: r.value, done: false };
-          }
-        },
-        function returnFn() {
-          done = true;
-          _closeIterator(outer);
-          return { value: undefined, done: true };
-        },
-      );
-    });
-  }
-
-  if (typeof Iproto.take !== "function") {
-    _installBuiltinMethod(Iproto, "take", 1, function (this: any, limit: any) {
-      const outer = _requireIteratorReceiver(this, "take");
-      let remaining = Number(limit);
-      if (!Number.isFinite(remaining)) remaining = Infinity;
-      remaining = Math.trunc(remaining);
-      if (remaining < 0) throw new RangeError("Iterator.prototype.take: limit must be non-negative");
-      let done = false;
-      return _makeHelperIterator(
-        function next() {
-          if (done || remaining <= 0) {
-            done = true;
-            return { value: undefined, done: true };
-          }
-          remaining--;
-          const r = outer.next();
-          if (r && r.done) {
-            done = true;
-            return { value: undefined, done: true };
-          }
-          return { value: r.value, done: false };
-        },
-        function returnFn() {
-          done = true;
-          _closeIterator(outer);
-          return { value: undefined, done: true };
-        },
-      );
-    });
-  }
-
-  if (typeof Iproto.drop !== "function") {
-    _installBuiltinMethod(Iproto, "drop", 1, function (this: any, limit: any) {
-      const outer = _requireIteratorReceiver(this, "drop");
-      let remaining = Number(limit);
-      if (!Number.isFinite(remaining)) remaining = Infinity;
-      remaining = Math.trunc(remaining);
-      if (remaining < 0) throw new RangeError("Iterator.prototype.drop: limit must be non-negative");
-      let done = false;
-      return _makeHelperIterator(
-        function next() {
-          if (done) return { value: undefined, done: true };
-          while (remaining > 0) {
-            remaining--;
-            const skipped = outer.next();
-            if (skipped && skipped.done) {
-              done = true;
-              return { value: undefined, done: true };
-            }
-          }
-          const r = outer.next();
-          if (r && r.done) {
-            done = true;
-            return { value: undefined, done: true };
-          }
-          return { value: r.value, done: false };
-        },
-        function returnFn() {
-          done = true;
-          _closeIterator(outer);
-          return { value: undefined, done: true };
-        },
-      );
-    });
-  }
-
-  if (typeof Iproto.toArray !== "function") {
-    _installBuiltinMethod(Iproto, "toArray", 0, function (this: any) {
-      const iter = _requireIteratorReceiver(this, "toArray");
-      const out: any[] = [];
-      for (;;) {
-        const r = iter.next();
-        if (r && r.done) return out;
-        out.push(r.value);
-      }
-    });
-  }
-
-  if (typeof Iproto.forEach !== "function") {
-    _installBuiltinMethod(Iproto, "forEach", 1, function (this: any, fn: any) {
-      const iter = _requireIteratorReceiver(this, "forEach");
-      if (typeof fn !== "function") {
-        throw new TypeError("Iterator.prototype.forEach: callback is not a function");
-      }
-      let counter = 0;
-      for (;;) {
-        const r = iter.next();
-        if (r && r.done) return undefined;
-        try {
-          fn(r.value, counter++);
-        } catch (e) {
-          _closeIterator(iter);
-          throw e;
-        }
-      }
-    });
-  }
-
-  if (typeof Iproto.some !== "function") {
-    _installBuiltinMethod(Iproto, "some", 1, function (this: any, predicate: any) {
-      const iter = _requireIteratorReceiver(this, "some");
-      if (typeof predicate !== "function") {
-        throw new TypeError("Iterator.prototype.some: predicate is not a function");
-      }
-      let counter = 0;
-      for (;;) {
-        const r = iter.next();
-        if (r && r.done) return false;
-        try {
-          if (predicate(r.value, counter++)) {
-            _closeIterator(iter);
-            return true;
-          }
-        } catch (e) {
-          _closeIterator(iter);
-          throw e;
-        }
-      }
-    });
-  }
-
-  if (typeof Iproto.every !== "function") {
-    _installBuiltinMethod(Iproto, "every", 1, function (this: any, predicate: any) {
-      const iter = _requireIteratorReceiver(this, "every");
-      if (typeof predicate !== "function") {
-        throw new TypeError("Iterator.prototype.every: predicate is not a function");
-      }
-      let counter = 0;
-      for (;;) {
-        const r = iter.next();
-        if (r && r.done) return true;
-        try {
-          if (!predicate(r.value, counter++)) {
-            _closeIterator(iter);
-            return false;
-          }
-        } catch (e) {
-          _closeIterator(iter);
-          throw e;
-        }
-      }
-    });
-  }
-
-  if (typeof Iproto.find !== "function") {
-    _installBuiltinMethod(Iproto, "find", 1, function (this: any, predicate: any) {
-      const iter = _requireIteratorReceiver(this, "find");
-      if (typeof predicate !== "function") {
-        throw new TypeError("Iterator.prototype.find: predicate is not a function");
-      }
-      let counter = 0;
-      for (;;) {
-        const r = iter.next();
-        if (r && r.done) return undefined;
-        try {
-          if (predicate(r.value, counter++)) {
-            _closeIterator(iter);
-            return r.value;
-          }
-        } catch (e) {
-          _closeIterator(iter);
-          throw e;
-        }
-      }
-    });
-  }
-
-  if (typeof Iproto.reduce !== "function") {
-    _installBuiltinMethod(Iproto, "reduce", 1, function (this: any, reducer: any, initial?: any) {
-      const iter = _requireIteratorReceiver(this, "reduce");
-      if (typeof reducer !== "function") {
-        throw new TypeError("Iterator.prototype.reduce: reducer is not a function");
-      }
-      let counter = 0;
-      let acc: any = initial;
-      if (arguments.length < 2) {
-        const first = iter.next();
-        if (first && first.done) {
-          throw new TypeError("Iterator.prototype.reduce: empty iterator with no initial value");
-        }
-        acc = first.value;
-        counter = 1;
-      }
-      for (;;) {
-        const r = iter.next();
-        if (r && r.done) return acc;
-        try {
-          acc = reducer(acc, r.value, counter++);
-        } catch (e) {
-          _closeIterator(iter);
-          throw e;
-        }
-      }
-    });
-  }
-
-  if (typeof I.zip !== "function") {
-    _installStaticHelper(I, "zip", 1, function zip(iterables: any, options?: any) {
-      if (!_isObject(iterables)) {
-        throw new TypeError("Iterator.zip: iterables must be an object");
-      }
-      const { mode, paddingOption } = _readZipOptions(options);
-      const iters: any[] = [];
-      let inputIter: any;
-      try {
-        inputIter = _getIterator(iterables);
-        for (;;) {
-          const next = _iteratorStepValue(inputIter);
-          if (next.done) break;
-          try {
-            iters.push(_getFlattenable(next.value, "reject-primitives"));
-          } catch (e) {
-            _closeIterator(inputIter);
-            _closeIterators(iters);
-            throw e;
-          }
-        }
-      } catch (e) {
-        _closeIterators(iters);
-        throw e;
-      }
-
-      const padding: any[] = [];
-      if (mode === "longest") {
-        if (paddingOption === undefined) {
-          for (let i = 0; i < iters.length; i++) padding.push(undefined);
-        } else {
-          let paddingIter: any;
-          try {
-            paddingIter = _getIterator(paddingOption);
-            let usingIterator = true;
-            for (let i = 0; i < iters.length; i++) {
-              if (usingIterator) {
-                const next = _iteratorStepValue(paddingIter);
-                if (next.done) {
-                  usingIterator = false;
-                } else {
-                  padding.push(next.value);
-                  continue;
-                }
-              }
-              padding.push(undefined);
-            }
-            if (usingIterator) {
-              paddingIter.return?.();
-            }
-          } catch (e) {
-            _closeIterators(iters);
-            throw e;
-          }
-        }
-      }
-
-      return _makeIteratorZip(iters, mode, padding, (results) => results.slice());
-    });
-  }
-
-  if (typeof I.zipKeyed !== "function") {
-    _installStaticHelper(I, "zipKeyed", 1, function zipKeyed(iterables: any, options?: any) {
-      if (!_isObject(iterables)) {
-        throw new TypeError("Iterator.zipKeyed: iterables must be an object");
-      }
-      const { mode, paddingOption } = _readZipOptions(options);
-      const keys: PropertyKey[] = [];
-      const iters: any[] = [];
-      try {
-        for (const key of Reflect.ownKeys(iterables)) {
-          const desc = Object.getOwnPropertyDescriptor(iterables, key);
-          if (desc && desc.enumerable) {
-            const value = iterables[key];
-            if (value !== undefined) {
-              keys.push(key);
-              iters.push(_getFlattenable(value, "reject-primitives"));
-            }
-          }
-        }
-      } catch (e) {
-        _closeIterators(iters);
-        throw e;
-      }
-
-      const padding: any[] = [];
-      if (mode === "longest") {
-        if (paddingOption === undefined) {
-          for (let i = 0; i < iters.length; i++) padding.push(undefined);
-        } else {
-          try {
-            for (const key of keys) {
-              padding.push(paddingOption[key]);
-            }
-          } catch (e) {
-            _closeIterators(iters);
-            throw e;
-          }
-        }
-      }
-
-      return _makeIteratorZip(iters, mode, padding, (results) => {
-        const out: any = Object.create(null);
-        for (let i = 0; i < keys.length; i++) {
-          Object.defineProperty(out, keys[i]!, {
-            value: results[i],
-            writable: true,
-            enumerable: true,
-            configurable: true,
-          });
-        }
-        return out;
-      });
-    });
-  }
-
-  if (typeof I.concat !== "function") {
-    _installStaticHelper(I, "concat", 0, function concat(...iterables: any[]) {
-      const records: { iterable: any; method: any }[] = [];
-      for (const iterable of iterables) {
-        if (!_isObject(iterable)) {
-          throw new TypeError("Iterator.concat: argument is not an object");
-        }
-        const method = iterable[Symbol.iterator];
-        if (method === undefined) {
-          throw new TypeError("Iterator.concat: argument is not iterable");
-        }
-        if (typeof method !== "function") {
-          throw new TypeError("Iterator.concat: argument is not iterable");
-        }
-        records.push({ iterable, method });
-      }
-      let idx = 0;
-      let current: any = null;
-      return _makeHelperIterator(
-        function next() {
-          while (true) {
-            if (current == null) {
-              if (idx >= records.length) return { value: undefined, done: true };
-              const record = records[idx++]!;
-              current = _getIteratorDirect(record.method.call(record.iterable));
-            }
-            let r: any;
-            try {
-              r = current.next();
-            } catch (e) {
-              current = null;
-              idx = records.length;
-              throw e;
-            }
-            if (!_isObject(r)) {
-              current = null;
-              idx = records.length;
-              throw new TypeError("Iterator result is not an object");
-            }
-            if (r.done) {
-              current = null;
-              continue;
-            }
-            return { value: r.value, done: false };
-          }
-        },
-        function returnFn() {
-          if (current != null) {
-            try {
-              current.return?.();
-            } catch {}
-          }
-          idx = iterables.length;
-          current = null;
-          return { value: undefined, done: true };
-        },
-      );
-    });
-  }
-
-  // #1718 S1 — Iterator.prototype.flatMap (TC39 iterator-helpers, ES2025).
-  // §27.1.4.x: for each value v of the underlying iterator, call
-  // mapper(v, counter); GetIteratorFlattenable(result, reject-primitives);
-  // yield every value of that inner iterator before advancing the outer.
-  // Hosts that lack the native helper (older V8 / Node) fall through to here;
-  // installing it on %Iterator.prototype% makes every helper-iterator and
-  // synthesized iterator (which inherit from Iproto, #1367) gain flatMap.
-  if (typeof Iproto.flatMap !== "function") {
-    Object.defineProperty(Iproto, "flatMap", {
-      value: function flatMap(mapper: any) {
-        // 1. RequireObjectCoercible(this) + this has [[next]] (it's an Iterator).
-        if (this == null || typeof this.next !== "function") {
-          throw new TypeError("Iterator.prototype.flatMap called on non-iterator");
-        }
-        // 2. mapper must be callable.
-        if (typeof mapper !== "function") {
-          throw new TypeError("Iterator.prototype.flatMap: mapper is not a function");
-        }
-        const outer: any = this;
-        let counter = 0;
-        let inner: any = null; // currently-open inner iterator, or null
-        let done = false;
-
-        function closeInner(): void {
-          if (inner != null) {
-            const innerIt = inner;
-            inner = null;
-            try {
-              innerIt.return?.();
-            } catch {}
-          }
-        }
-
-        return _makeHelperIterator(
-          function next() {
-            if (done) return { value: undefined, done: true };
-            while (true) {
-              if (inner == null) {
-                // Advance the outer iterator.
-                let outerRes: any;
-                try {
-                  outerRes = outer.next();
-                } catch (e) {
-                  done = true;
-                  throw e;
-                }
-                if (outerRes && outerRes.done) {
-                  done = true;
-                  return { value: undefined, done: true };
-                }
-                // mapped = mapper(value, counter); IfAbruptCloseIterator(outer).
-                let mapped: any;
-                try {
-                  mapped = mapper(outerRes.value, counter);
-                } catch (e) {
-                  done = true;
-                  try {
-                    outer.return?.();
-                  } catch {}
-                  throw e;
-                }
-                counter++;
-                try {
-                  inner = _getFlattenable(mapped, "reject-primitives");
-                } catch (e) {
-                  done = true;
-                  try {
-                    outer.return?.();
-                  } catch {}
-                  throw e;
-                }
-              }
-              // Pull from the inner iterator.
-              let innerRes: any;
-              try {
-                innerRes = inner.next();
-              } catch (e) {
-                done = true;
-                inner = null;
-                try {
-                  outer.return?.();
-                } catch {}
-                throw e;
-              }
-              if (innerRes && innerRes.done) {
-                inner = null;
-                continue; // inner exhausted — advance outer
-              }
-              return { value: innerRes.value, done: false };
-            }
-          },
-          function returnFn() {
-            done = true;
-            closeInner();
-            try {
-              outer.return?.();
-            } catch {}
-            return { value: undefined, done: true };
-          },
-        );
-      },
-      writable: true,
-      configurable: true,
-    });
-  }
-}
 
 /** Tracks WasmGC struct objects that have been frozen via Object.freeze. */
 const _wasmFrozenObjs = new WeakSet<object>();
@@ -15280,50 +13998,6 @@ assert._isSameValue = isSameValue;
   }
 }
 
-/**
- * Build string constants object for the "string_constants" import namespace.
- * Each string pool entry becomes a WebAssembly.Global keyed by the literal text.
- */
-export function buildStringConstants(stringPool: string[] = []): Record<string, WebAssembly.Global> {
-  // Use a null-prototype object so inherited names like "hasOwnProperty" /
-  // "toString" / "constructor" from Object.prototype don't shadow real pool
-  // entries via the `s in constants` duplicate check.
-  const constants: Record<string, WebAssembly.Global> = Object.create(null);
-  for (const s of stringPool) {
-    // (#2880) Literals with a lone surrogate are imported from the
-    // `string_constants16` namespace (hex-keyed) instead — see
-    // buildStringConstants16. Skip them here so their lossy UTF-8 text isn't
-    // also installed under a wrong key.
-    if (hasLoneSurrogate(s)) continue;
-    if (!(s in constants)) {
-      constants[s] = new WebAssembly.Global({ value: "externref", mutable: false }, s);
-    }
-  }
-  return constants;
-}
-
-/**
- * (#2880) Surrogate-safe sibling of {@link buildStringConstants}. A wasm import
- * field name must be valid UTF-8, so a string literal containing a lone
- * (unpaired) surrogate cannot be keyed by its own text. The compiler instead
- * imports such constants from the `string_constants16` namespace keyed by the
- * hex of their UTF-16 code units (ASCII). This mirrors that keying: the lookup
- * key is the hex, the global's VALUE is the real JS string (externref holds any
- * string, lone surrogates included). Surrogate-free literals are never routed
- * here — they stay in the plain `string_constants` namespace, unchanged.
- */
-export function buildStringConstants16(stringPool: string[] = []): Record<string, WebAssembly.Global> {
-  const constants: Record<string, WebAssembly.Global> = Object.create(null);
-  for (const s of stringPool) {
-    if (!hasLoneSurrogate(s)) continue;
-    const key = hexCodeUnits(s);
-    if (!(key in constants)) {
-      constants[key] = new WebAssembly.Global({ value: "externref", mutable: false }, s);
-    }
-  }
-  return constants;
-}
-
 /** Check a manifest against a policy blocklist before instantiation.
  *  Returns an array of violated import keys (empty if all clear). */
 export function checkPolicy(manifest: ImportDescriptor[], policy: ImportPolicy): string[] {
@@ -15465,304 +14139,6 @@ function wrapWithContainment(
 
   // Default: return original
   return fn;
-}
-
-/**
- * Build a WASI polyfill for running WASI-compiled modules in JS environments.
- * Routes fd_write(fd=1) to console.log, fd_write(fd=2) to console.error,
- * and proc_exit to process.exit (Node) or throw (browser).
- *
- * Usage:
- *   const wasi = buildWasiPolyfill();
- *   const { instance } = await WebAssembly.instantiate(binary, {
- *     wasi_snapshot_preview1: wasi,
- *     env: wasi.envImports,
- *   });
- *   wasi.setMemory(instance.exports.memory as WebAssembly.Memory);
- *   (instance.exports._start as Function)();
- *
- * #1482: The polyfill now exposes `envImports.__wasi_env_get_str` for the
- * `process.env.X` fast path under `--target wasi`, plus `environ_sizes_get`
- * and `environ_get` shims (memory-writing) for true WASI hosts. The defaults
- * read from Node's `process.env`; pass `{ env: {...} }` to override.
- */
-export function buildWasiPolyfill(options?: { env?: Record<string, string | undefined> }): {
-  fd_write: (fd: number, iovs: number, iovs_len: number, nwritten: number) => number;
-  fd_read: (fd: number, iovs: number, iovs_len: number, nread: number) => number;
-  proc_exit: (code: number) => void;
-  poll_oneoff: (in_ptr: number, out_ptr: number, nsubs: number, nevents_out: number) => number;
-  fd_fdstat_set_flags: (fd: number, flags: number) => number;
-  environ_sizes_get: (countPtr: number, bufSizePtr: number) => number;
-  environ_get: (envPtrsPtr: number, envBufPtr: number) => number;
-  clock_time_get: (clockid: number, precision: bigint, out_ptr: number) => number;
-  setMemory: (mem: WebAssembly.Memory) => void;
-  setStdin: (data: Uint8Array | string) => void;
-  envImports: Record<string, Function>;
-} {
-  let memory: WebAssembly.Memory | undefined;
-  // Partial line buffer per fd for data not ending in newline
-  const lineBuffers: Record<number, string> = {};
-  // (#1483) Monotonic baseline so CLOCK_MONOTONIC values start near zero and
-  // never go backwards within a single instance lifetime.
-  const monotonicStartNs = (() => {
-    const perf = typeof performance !== "undefined" && typeof performance.now === "function" ? performance : undefined;
-    return perf ? BigInt(Math.round(perf.now() * 1_000_000)) : BigInt(Date.now()) * 1_000_000n;
-  })();
-  // Buffered stdin bytes; consumed by fd_read until EOF (length 0).
-  // Tests/harnesses can preload bytes via setStdin().
-  let stdinBuf: Uint8Array = new Uint8Array(0);
-  let stdinPos = 0;
-
-  // #1482: source of environment data. Caller-supplied dict wins; otherwise
-  // default to Node's `process.env` (browser → empty dict).
-  const envSource: Record<string, string | undefined> =
-    options?.env ??
-    (typeof process !== "undefined" && process.env ? (process.env as Record<string, string | undefined>) : {});
-
-  const polyfill = {
-    setMemory(mem: WebAssembly.Memory) {
-      memory = mem;
-    },
-
-    /** Preload stdin bytes for the next sequence of fd_read calls. */
-    setStdin(data: Uint8Array | string) {
-      stdinBuf = typeof data === "string" ? new TextEncoder().encode(data) : data;
-      stdinPos = 0;
-    },
-
-    /**
-     * Minimal fd_read for fd=0 (stdin). Reads from the preloaded buffer
-     * (see setStdin); returns 0 bytes (EOF) once exhausted. fd != 0 yields
-     * EBADF-like behavior by writing nread=0 and returning 0.
-     */
-    fd_read(fd: number, iovs: number, iovs_len: number, nread: number): number {
-      if (!memory) return -1;
-      const view = new DataView(memory.buffer);
-      let totalRead = 0;
-
-      if (fd === 0) {
-        for (let i = 0; i < iovs_len; i++) {
-          const ptr = view.getUint32(iovs + i * 8, true);
-          const len = view.getUint32(iovs + i * 8 + 4, true);
-          if (len === 0) continue;
-          const remaining = stdinBuf.length - stdinPos;
-          if (remaining <= 0) break;
-          const take = Math.min(len, remaining);
-          const dest = new Uint8Array(memory.buffer, ptr, take);
-          dest.set(stdinBuf.subarray(stdinPos, stdinPos + take));
-          stdinPos += take;
-          totalRead += take;
-          if (take < len) break; // partial fill = drained
-        }
-      }
-
-      view.setUint32(nread, totalRead, true);
-      return 0; // __WASI_ERRNO_SUCCESS
-    },
-
-    fd_write(fd: number, iovs: number, iovs_len: number, nwritten: number): number {
-      if (!memory) return -1; // EBADF-ish: memory not set
-
-      const view = new DataView(memory.buffer);
-      let totalWritten = 0;
-
-      for (let i = 0; i < iovs_len; i++) {
-        const ptr = view.getUint32(iovs + i * 8, true);
-        const len = view.getUint32(iovs + i * 8 + 4, true);
-        const bytes = new Uint8Array(memory.buffer, ptr, len);
-        const text = new TextDecoder().decode(bytes);
-
-        // Buffer partial lines; flush on newline
-        const buf = (lineBuffers[fd] || "") + text;
-        const lines = buf.split("\n");
-        // Last element is the incomplete line (or "" if text ended with \n)
-        lineBuffers[fd] = lines.pop()!;
-        const writer = fd === 2 ? console.error : console.log;
-        for (const line of lines) {
-          writer(line);
-        }
-
-        totalWritten += len;
-      }
-
-      // Write total bytes written
-      view.setUint32(nwritten, totalWritten, true);
-      return 0; // __WASI_ERRNO_SUCCESS
-    },
-
-    proc_exit(code: number): void {
-      if (typeof process !== "undefined" && typeof process.exit === "function") {
-        process.exit(code);
-      }
-      throw new Error(`WASI proc_exit(${code})`);
-    },
-
-    // #1484 / #2632 — poll_oneoff shim for vitest-driven tests.
-    //
-    // Real wasmtime semantics: read `nsubs` subscription_t (48 bytes each) from
-    // `in_ptr`, suspend until the earliest event fires, then write the firing
-    // event_t (32 bytes each) records to `out_ptr` and the count to
-    // `nevents_out`. Returns 0 (__WASI_ERRNO_SUCCESS).
-    //
-    // We support two shapes:
-    //   - Phase 1 single CLOCK subscription (`__wasi_sleep_ms`): acknowledge it
-    //     synchronously (no real sleep — tests run instantly), report 1 clock
-    //     event fired.
-    //   - #2632 Phase 2 multi-sub (fd0 FD_READ + optional CLOCK): if fd0 has
-    //     buffered/preloaded stdin remaining, report the FD_READ event as fired
-    //     (the reactor then fd_reads it); otherwise report the CLOCK event
-    //     (timer due / timeout). This lets vitest exercise the fd-readiness
-    //     reactor without a real OS poll.
-    poll_oneoff(in_ptr: number, out_ptr: number, nsubs: number, nevents_out: number): number {
-      if (!memory) return -1;
-      const view = new DataView(memory.buffer);
-
-      // Decode subscriptions: tag (u8) @ off+8 → 0=CLOCK, 1=FD_READ; for
-      // FD_READ the fd (u32) is @ off+16.
-      type Sub = { type: number; fd: number; userdata: bigint };
-      const subs: Sub[] = [];
-      for (let s = 0; s < nsubs; s++) {
-        const off = in_ptr + s * 48;
-        const userdata = view.getBigUint64(off, true);
-        const tag = view.getUint8(off + 8);
-        const fd = tag === 1 ? view.getUint32(off + 16, true) : -1;
-        subs.push({ type: tag, fd, userdata });
-      }
-
-      // Does fd0 have stdin remaining to read? (preloaded via setStdin / unread)
-      const fd0Readable = stdinBuf.length - stdinPos > 0;
-      const fd0Sub = subs.find((x) => x.type === 1 && x.fd === 0);
-      const clockSub = subs.find((x) => x.type === 0);
-
-      // Pick the firing event(s): FD_READ wins when fd0 is readable; else the
-      // clock fires (timeout elapses instantly in the polyfill).
-      const fired: Sub[] = [];
-      if (fd0Sub && fd0Readable) {
-        fired.push(fd0Sub);
-      } else if (clockSub) {
-        fired.push(clockSub);
-      } else if (fd0Sub) {
-        // No clock and fd0 not readable — report fd0 anyway so the reactor's
-        // 0-byte (EOF) read path ends the subscription instead of hanging.
-        fired.push(fd0Sub);
-      } else if (subs.length > 0) {
-        fired.push(subs[0]!);
-      }
-
-      // Write event_t records: [0..7] userdata, [8..9] errno=0, [10] type.
-      let n = 0;
-      for (const ev of fired) {
-        const eoff = out_ptr + n * 32;
-        for (let i = 0; i < 32; i++) view.setUint8(eoff + i, 0);
-        view.setBigUint64(eoff, ev.userdata, true);
-        view.setUint16(eoff + 8, 0, true); // errno
-        view.setUint8(eoff + 10, ev.type); // type: 0=CLOCK, 1=FD_READ
-        n++;
-      }
-      view.setUint32(nevents_out, n, true);
-      return 0;
-    },
-
-    /**
-     * #2632 Phase 2 — fd_fdstat_set_flags(fd, flags) -> errno. The reactor calls
-     * this to put fd 0 in non-blocking mode. The polyfill's fd_read is already
-     * non-blocking (returns immediately), so this is a no-op acknowledgement.
-     */
-    fd_fdstat_set_flags(_fd: number, _flags: number): number {
-      return 0;
-    },
-
-    // #1482: WASI environ_sizes_get — report `[count, total_buf_bytes]`. The
-    // buffer layout we report (when environ_get fires) is `KEY=VALUE\0`
-    // repeated, UTF-8 encoded. We compute it from `envSource` on each call;
-    // the cost is negligible for typical env sizes and avoids stale results
-    // when callers mutate the source between invocations.
-    environ_sizes_get(countPtr: number, bufSizePtr: number): number {
-      if (!memory) return -1;
-      const entries = Object.entries(envSource).filter(([, v]) => v !== undefined) as [string, string][];
-      const enc = new TextEncoder();
-      let bufBytes = 0;
-      for (const [k, v] of entries) {
-        bufBytes += enc.encode(`${k}=${v}`).length + 1; // +1 for NUL terminator
-      }
-      const view = new DataView(memory.buffer);
-      view.setUint32(countPtr, entries.length, true);
-      view.setUint32(bufSizePtr, bufBytes, true);
-      return 0;
-    },
-
-    // #1482: WASI environ_get — write the env pointer table at `envPtrsPtr`
-    // and the `KEY=VALUE\0...` buffer at `envBufPtr`. Iteration order MUST
-    // match what environ_sizes_get reported, otherwise the guest's allocator
-    // will mis-size the buffer.
-    environ_get(envPtrsPtr: number, envBufPtr: number): number {
-      if (!memory) return -1;
-      const entries = Object.entries(envSource).filter(([, v]) => v !== undefined) as [string, string][];
-      const view = new DataView(memory.buffer);
-      const mem = new Uint8Array(memory.buffer);
-      const enc = new TextEncoder();
-      let cursor = envBufPtr;
-      for (let i = 0; i < entries.length; i++) {
-        const [k, v] = entries[i]!;
-        view.setUint32(envPtrsPtr + i * 4, cursor, true);
-        const bytes = enc.encode(`${k}=${v}`);
-        mem.set(bytes, cursor);
-        cursor += bytes.length;
-        mem[cursor++] = 0; // NUL terminator
-      }
-      return 0;
-    },
-
-    /**
-     * #1482: env-namespace host imports for compiled modules that use
-     * `process.env.X` under `--target wasi`. Wire as `{ env: wasi.envImports }`
-     * alongside `wasi_snapshot_preview1: wasi` when instantiating.
-     *
-     * `__wasi_env_get_str(key)` is the JS-polyfill fast path — it returns
-     * the value (or `undefined`) directly as a JS string, sidestepping the
-     * memory marshalling of `environ_get`. The Wasm side type signature is
-     * `(externref) -> externref`.
-     */
-    envImports: {
-      __wasi_env_get_str(key: unknown): string | undefined {
-        if (typeof key !== "string") return undefined;
-        const v = envSource[key];
-        return v === undefined ? undefined : v;
-      },
-    } as Record<string, Function>,
-
-    /**
-     * (#1483) clock_time_get(clockid, precision, out_ptr) -> errno
-     *
-     * Writes the current time in nanoseconds as a little-endian u64 to
-     * out_ptr in the module's linear memory. Supports:
-     *   - CLOCK_REALTIME   (0) → Date.now() (wall-clock ms → ns)
-     *   - CLOCK_MONOTONIC  (1) → performance.now() (sub-ms, monotonic)
-     *
-     * `precision` is advisory — we always report ns granularity from
-     * whichever JS clock is available.
-     */
-    clock_time_get(clockid: number, _precision: bigint, out_ptr: number): number {
-      if (!memory) return 28; // EINVAL — memory not set
-      let nowNs: bigint;
-      if (clockid === 1) {
-        // CLOCK_MONOTONIC — sub-ms via performance.now if available.
-        const perf =
-          typeof performance !== "undefined" && typeof performance.now === "function" ? performance : undefined;
-        nowNs = perf ? BigInt(Math.round(perf.now() * 1_000_000)) : BigInt(Date.now()) * 1_000_000n;
-        nowNs -= monotonicStartNs;
-        if (nowNs < 0n) nowNs = 0n;
-      } else {
-        // CLOCK_REALTIME (0) and unknown clock ids fall through to wall-clock ms.
-        nowNs = BigInt(Date.now()) * 1_000_000n;
-      }
-      const view = new DataView(memory.buffer);
-      view.setBigUint64(out_ptr, nowNs, true);
-      return 0;
-    },
-  };
-
-  return polyfill;
 }
 
 /** Build the WebAssembly import object from a closed manifest */
